Copy and map kernel to higher half

+++ lib/chos-boot-defs/src/lib.rs
@@ -1,12 +1,12 @@
 #![no_std]
 
 pub mod phys {
-    pub const KERNEL_DATA_BASE: usize = 0x0100_0000;
+    pub const KERNEL_DATA_BASE: u64 = 0x0100_0000;
 }
 
 pub mod virt {
-    pub const KERNEL_CODE_BASE: usize = 0xffff_8000_4000_0000;
-    pub const KERNEL_PT_BASE: usize =   0xffff_8000_8000_0000;
+    pub const KERNEL_CODE_BASE: u64 = 0xffff_8000_4000_0000;
+    pub const KERNEL_PT_BASE: u64 =   0xffff_8000_8000_0000;
 }
 
 pub enum BootMemoryType {
